pmg: disable postgres backup and replication services
If installed manually on a running system, or via our ISO these
services do not get enabled.

having them enabled leads to error messages in the journal, due to the
pg_hba.conf.in template in pmg-api being out-of-date.
In any case these should not be enabled by default - see:
/usr/share/doc/postgresql-common/README.systemd

I currently suspect that the issue with the template here is that
firstboot services run with it - and systemd enables the
template-units (instead of the instaniated ones) e.g.:
/etc/systemd/system/postgresql@.service.wants/pg_receivewal@.service

which then get started because of postgresql@15.service being enabled.

With the presets added the units remain disabled.

reported in our community forum:
https://forum.proxmox.com/threads/.130704/

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>